+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2024-01-01
+### Added
+- In `TimeDisplay` added also the year
+
 ## [v15.11.0] - 2023-12-31
 ### Added
 - `TimeDisplay` on the right side of `NavMenu`